pyproject.toml
[project]
name = "llmops"
version = "0.1.0"
description = "Multi-document RAG chat application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "faiss-cpu>=1.9.0",
    "tiktoken>=0.7.0",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
    "docx2txt>=0.8",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
]
